fix: 修复 Decimal 序列化为字符串导致页面崩溃的问题
- 后端: 输出 Schema 的 Decimal 字段改为 float，确保 JSON 序列化为数字
- 后端: accounts router 中 compute_balance 返回值用 float() 显式转换
- 前端: 所有 .toFixed() 调用前加 Number() 防御性转换
- 新增 PROJECT.md 项目文档

+++ backend/app/schemas/account.py
@@ -28,10 +28,10 @@ class AccountOut(BaseModel):
     account_type: str
     icon: str | None
     color: str | None
-    initial_balance: Decimal
+    initial_balance: float
     is_active: bool
     sort_order: int
     created_at: datetime
-    balance: Decimal | None = None
+    balance: float | None = None
 
     model_config = {"from_attributes": True}
